[TASK] Migrate extension commands to Symfony
The Extbase command controller from Extension Manager is the only
CommandController used in TYPO3 Core (apart from extbase:help), all
others have been migrated to symfony/console commands.

This however is now only possible due to the recent migration to define
these commands as "non-schedulable" so they do not show up in the
scheduler as a possible command to be called repeatedly.

For the "extensionmanager:extension:dumpclassloadinginformation"
call, the functionality is moved to EXT:core as this is not related to any
functionality in ExtensionManager, as it does not need any information
of this. This command has been made available under "dumpautoload"
to get closer on the naming scheme of native composer.

"install" and "uninstall" however rely on the DI concept of Extbase and
Extensionmanager to inject necessary components and is tightly
intertwined. These are now provided as "extension:activate" and
"extension:deactivate" respectively to convey their purpose and especially
show that the latter does not remove any data or code.

Releases: master
Resolves: #85996
Change-Id: I8025d8ba8f577c46815996bd3c7d13d17dec268a
Reviewed-on: https://review.typo3.org/58055
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>