[BUGFIX] Adapt view related install tool details
With the install tool move to modals, various details on CSS
level broke: The modal HTML is on top level in backend and thus
not within the content-iframe, so styles defined in install.css
loaded in the iframe do not kick in. Since the standalone install
tool does not use iframes, there are various differences on view
level between standalone and embedded-in-backend version, usually
with the embedded version looking more ugly than standalone.

The install.css is not very different from backend.css anyway,
so we resolve various details with the patch, kick install.css
entirely and rely soley on backend.css, even for the installer.

Details:
* <hr> shows a border again in modals in embedded, broken since
  modal patch
* image processing view has borders and other details again in
  embedded verision, broken since modal patch
* removed 'fixed save button location' for 'all configuration' code.
  This broke with the modal patch, the code does not do anything
  useful at the moment. A solution needs a change of the modal js
  to render the buttons to modal-footer instead. This is too complex
  for this patch and may be done with another patch.
* A couple of classes to easily limit styles to installer and
  maintenance parts
* Remove left overs from old install-tool menu approach that
  have not been cleanup up, yet.

Resolves: #86245
Related: #84772
Releases: master
Change-Id: I4d21677331112a48f84b0cf48a574999128a15b7
Reviewed-on: https://review.typo3.org/58269
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>