[TASK] Improve UI of „Upgrade Wizard“
Loading the list of available upgrade wizards may take some time. To give
a hint what is currently done, the progress bar now shows a text that
describes the currently processed step. On initial load, the progress
bar of the "done" wizards is rendered as well.

When a upgrade wizard is about to get executed, the viewport scrolls to
the correct position now to attract attention.

Resolves: #86278
Releases: master
Change-Id: Ie9c7d4cdf7e125374f8560fca559e7821b204110
Reviewed-on: https://review.typo3.org/58294
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>