[BUGFIX] Improve performance of extension scanner
This patch adds a request queue for the massive amount of AJAX calls
in the extension scanner. This patch mitigates the problem of massive
server load in case the amount of files to be scanned is very high.

This patch mitigates the problem by allowing a maximum of 10 concurrent
requests. After the release of TYPO3 v9 LTS, this scanner module should
be refactored further.

Resolves: #86436
Releases: master
Change-Id: I824dfb74aae46d6b00690e7ca6553305c4221f0f
Reviewed-on: https://review.typo3.org/58510
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>