[BUGFIX] Fix decrementing of AJAX requests in extension scanner
The queueing of AJAX requests in the extension scanner has some flaws
which causes the scanner to refuse another scan request.

The counter of pending requests is now always decremented, not only if
the amount of requests equals the threshold. Also, it's now checked if
there is any item in the request queue before shifting.

Instead of overriding the `success` and `error` callback, only `complete`
is now overridden, as this callback is executed on any return state.

To make the queueing part better maintainable and reusable, the code has
been moved into a separate module.

Resolves: #86523
Related: #86436
Releases: master
Change-Id: Ifa26ad7ce3bd81497c43b8606b902d467bdd1dda
Reviewed-on: https://review.typo3.org/58539
Reviewed-by: Richard Haeser <richard@maxserv.com>
Tested-by: Richard Haeser <richard@maxserv.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>