[BUGFIX] Create default site configuration on installation
When a new TYPO3 installation is set up, the user can choose
to "create a first site" (page with sys_template). In this step,
it should be important that a site configuration should be created
as well, in order to use site handling right out-of-the-box.

Therefore, this step in the installer now also creates a site
configuration (identifier "main") with one language english
to get started.

Resolves: #87383
Releases: master, 9.5
Change-Id: Ib8ef27c076d8ed24bfeb015aea52bdcf4ec0432c
Reviewed-on: https://review.typo3.org/59391
Reviewed-by: Steffen Dietrich <Teilzeitphilosoph@gmx.de>
Tested-by: Steffen Dietrich <Teilzeitphilosoph@gmx.de>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>