[TASK] Improve ClassSchema methods api
This patch introduces a new Method class to have an
easier and more stable api to get information about class
methods from ClassSchema objects.

The reflection data fetched and stored by/in the ClassSchema
objects is no longer publicly available as an array. Instead,
Method objects are returned for the following calls:

- getMethod
- getMethods
- getInjectMethods

Also it removes the following method:

- getConstructorArguments

The following constants are marked as private:

- ClassSchema::MODELTYPE_ENTITY
- ClassSchema::MODELTYPE_VALUEOBJECT

Releases: master
Resolves: #87427
Change-Id: I38e8a7287896e9303fad1fa8f8f9182895365adc
Reviewed-on: https://review.typo3.org/59422
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Tested-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: André Schließer <andy.schliesser@gmail.com>
Tested-by: André Schließer <andy.schliesser@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>