[!!!][TASK] Remove TCA for sys_history
TCA for sys_history is not necessary at all anymore, and
can be removed.

It was previously used for creating an Extbase Domain Model
for the BELog module created in TYPO3 6.0/6.1.

However, in 9.0 sys_history was decoupled from sys_log,
whereas the Domain Model and the usages for TCA was
not necessary anymore.

The previously "generated" database fields are re-added
manually in ext_tables.sql. The field "pid" which was necessary
for a TCA table is now dropped.

Resolves: #87936
Releases: master
Change-Id: Ia541e9a5c7d85f25b6ff3ceb7e3ee311300762e6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60273
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Alexander Schnitzler <typo3@alexanderschnitzler.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Alexander Schnitzler <typo3@alexanderschnitzler.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>