[BUGFIX] Make wizard SeparateSysHistoryFromSysLogUpdate repeatable
To prevent memory issues with huge sys_history tables and to prevent
timeouts when handling large amounts of data the wizard processes small
batches of records at a time, keeps track of its progress in the registry
and this way allows repeated runs without losing data.
If both the registry and sys_history share the same connection a database
transaction is used to make sure that modifying data and progress
tracking are in sync.

Resolves: #88487
Releases: master, 9.5
Change-Id: I7d323ac3dae1d416f5883b2e52f94ee50bd66327
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63577
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Susanne Moog <look@susi.dev>