[FEATURE] Provide feature flag for new translation server's export
The feature switch `betaTranslationServer` makes it possible for users
to test and use the new translation server.

The integration is currently work in progress but will be finished before
the LTS release of version 10.
Once the work has been stabilized and tested well,
the feature flag will be removed for 10 and backported for 9.

Resolves: #89526
Releases: master
Change-Id: Ic88e4b6568dbe53cd88e63a8a1dd39787219e9e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62127
Tested-by: Joerg Kummer <typo3@enobe.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Joerg Kummer <typo3@enobe.de>
Reviewed-by: Benni Mack <benni@typo3.org>