[BUGFIX] PopulatePagesSlugs should respect tx_realurl_pathsegment field
For upgrading from an old versions where EXT:realurl was used
the PopulatePagesSlugs upgradeWizard is in place.
It now also respects the tx_realurl_pathsegment field in generatorOptions.

The workflow is now, that it takes the page:tx_realurl_pathsegment field
and if empty falls back to page:title.

Resolves: #89069
Releases: master, 9.5
Change-Id: Ia9c6367d46713a8a1a609ef13c6bbd8878ec6bed
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62669
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>