[BUGFIX] Resolve nested public path inside ext:install DefaultFactory
This patch provides support for nested public paths in the TYPO3
installer. The DefaultFactory iterates backwards through the components
of the path to build a correct array.
To keep the source code as simple as possible, the substructure is
already built and passed to the function. Thus references in the code
can be avoided.

Resolves: #90690
Releases: master
Change-Id: Ib3bb3f265f0e0dc90a56fa2275f12e6cc1e602b1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63640
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Rico Drexler <rico.drexler@gmx.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>