[BUGFIX] Fix various issues in "View Upgrade Docs"
The search field was initialized twice which broke the clearable
feature. Since we don't like redundancy we omit a second initialization
and better do it right the first time.

If the search field is emptied after querying the docs, the original
state is recovered. All panels without exception are closed again.

When panels get collapsed, only expanded panels are queried which
improves the subjective performance of the process.

An unavailable function `.contains` was used to filter changelogs by
tags. Searching for a colon is now done via
`String.prototype.includes()`.

Any search action is now debounced to avoid massive DOM manipulations
due to excessive typing in the search field.

Resolves: #90737
Resolves: #85453
Resolves: #90738
Releases: master
Change-Id: Ia3b4f2665c95e9c7e891feec73b57b869279d22b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63688
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>