[BUGFIX] Fix type error in log upgrade wizard
During the update, the value of the field log_data is
unserialized. The expected value is an array, but it was
never validated. If the field contains an invalid value,
unserialize() returns a boolean false.

Now the result is type casted to an array to ensure the
correct type.

Resolves: #91042
Releases: 10.4, 9.5
Change-Id: I2959da0ae5b74f0377285755175da447676ead68
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66114
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>