[BUGFIX] Do not deprecate $GLOBALS[TYPO3_REQUEST]
The global object $GLOBALS[TYPO3_REQUEST] holding the
current PSR-7 request object was introduced in TYPO3 v9.2.

However, it was also marked as deprecated as we thought we
were able to remove all usages again by the end of TYPO3 v9.5
development, which shows that we have more problems in various
areas like hooks and Extbase where we heavily rely on this object.

For this reason, it is kept but the original Feature RST still contains
the information that it is considered bad practice.

Removing the deprecation will result in better result for the
ExtensionScanner.

Resolves: #91347
Releases: master, 9.5
Change-Id: I97bb16cf7f4e7149c5c3a3528a015701f60c2628
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64364
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>