[BUGFIX] Fix SMTP encryption migration when plaintext was used before
The SMTP SSL/TLS migration introduced in #91070 does not take the case
into account when no SMTP encryption was used at all (that means insecure
plaintext authentication). This could be configured by specifying an empty
string for `transport_smtp_encrypt` in TYPO3 v9.

We do now check for this third option and adapt the migration
to set the value to false, which means symfony/mailer will allow
connection without encryption. Note: symfony/mailer will still try to
start a STARTTLS connection if the server supports that capability.
(That is now default in symfony/mailer and can't be deactivated)

We also fix the default configuration of transport_smtp_encrypt
to be a boolean value. The setting was switched to boolean in #90295
but was forgotten to be adapted here.

Releases: master
Resolves: #91391
Related: #91070
Related: #90295
Change-Id: I16f0f19cf91b92b3a252d2a52c7226dd0eb23296
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64490
Reviewed-by: Josef Glatz <josefglatz@gmail.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Josef Glatz <josefglatz@gmail.com>
Tested-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>