[BUGFIX] Report database migration errors in upgrade wizard
When errors occur during the upgradeWizardsBlockingDatabaseExecute
ajax action, we now report SQL errors that occurred to the
integrator and pause further processing of upgrade wizards as some
wizards depend on an updated database structure.

The integrator may then decide to retry the migration
or to continue despite of errors.

An example:
If `slug` can not be added to the `pages` table, the
PopulatePageSlugs wizard will fail in updateNeccessary().
This causes all upgrades to be blocked but the cause
of the problem – why pages.slug is missing (e.g. because row size
is too large) – is not presented to the integrator.

Rather report these errors now and stop further processing
(which likely leads to a followup-error which is then hard to debug).

Releases: master, 10.4
Resolves: #91423
Change-Id: I526451617b0745a16153f7fcefaf4694f4a69481
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71408
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>