[FEATURE] Allow system maintainer to mute disable_functions error
Add a configuration option to customize the environment check in the
Install Tool for a list of `disable_functions` which are then reported
as warning instead of error.

With this new configuration option, a system maintainer is able to add
function names to this list. Thus, the Install Tool and monitoring
systems based on the environment check will not trigger errors for
these functions anymore. Since a warning will still be emitted,
the fact of a "special configuration" itself will not be lost.

Resolves: #91783
Releases: main
Change-Id: I50332c53fceda04d003ab0b67bf0a640c5c82866
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65032
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>