[TASK] Deprecate last arguments of wrapClickMenuOnIcon()
BackendUtility::wrapClickMenuOnIcon() has a boolean flag
to let the method return an array with tag parameters
instead of a fully build HTML tag as string.

As this are two completely different things and cause
problems when analysing return types it should not be
done in the same method.

Calling BackendUtility::wrapClickMenuOnIcon() with the
7th and last argument $returnTagParameters set to true
has been deprecated alongside the 5th and 6th arguments
that are already unused.

Resolves: #92583
Releases: master
Change-Id: Ia536aa3a52085e4ebe7956997b505047340102e7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66161
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>