[!!!][TASK] Remove unused possibility for user-based authentication timeout field
The AbstractUserAuthentication property "auth_timeout_field"
was used in the past (until TYPO3 8.0) to be filled for backend purposes
with "$GLOBALS['TYPO3_CONF_VARS']['BE']['sessionTimeout']"
and for backend with the lifetime field.

This field was not properly filled since TYPO3 v8.0, see
issue #68890 for details.

As the field had a dual-use but now is unused, it is properly
removed as TYPO3 Core never implemented this on a per-userrecord-basis
but handles this via the sessionTimeout propery now.

Resolves: #92802
Related: #68890
Releases: master
Change-Id: I760b50a292b93229bbebffac08e11393fe53393f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66595
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>