[!!!][TASK] Drop workspace additional services
Back in 6.2 at ExtJS times, a feature has been introduced
to the workspace module to extend the shown 'data grid'
with custom columns. This implementation has been done for
one specific customer who nowadays does not use the feature
anymore.
The feature broke when the transition to a native JavaScript
implementation has been done in v8. There is not a single bug
report this feature broke in forge, and it is hard to resurrect
it with the new client side implementation.
The patch drops related code from the extension. This gives
core development more freedom to improve the workspace module
with other patches.

Change-Id: Ie66b172484cdd08de06e019aa004055975948e85
Resolves: #92838
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66632
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Haupt <mail@danielhaupt.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Daniel Haupt <mail@danielhaupt.de>
Reviewed-by: Benni Mack <benni@typo3.org>