[BUGFIX] Ensure sys_log.level is varchar
The update wizard should make sure that the
level field is migrated from integer to varchar
otherwise a BE login is not possible.

The Database Schema Upgrade can handle new fields
properly, but for existing installations the existing
"level" field is an integer and would not allow
new log messages to the database. For this reason,
this workaround is added.

Resolves: #94460
Releases: master
Change-Id: I0ccb43239970030b50c2e59e288b0d830dbc2024
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69684
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>