[TASK] Deprecate GeneralUtility::shortMD5()
The method GeneralUtility::shortMD5() generates an MD5 hash and trims it
to a configurable length, where 10 characters is the default. This
increases the probability of conflicts whereever this method is used,
either as hash stored in the database or as id decorator in DOM.

Due to the issues this method brings, it is marked as deprecated now.

Resolves: #94684
Releases: master
Change-Id: I599211196da8ffd737643d29ed68dc6d0c0f2ae9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70176
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>