[TASK] Set explicitADmode to explicitAllow in DefaultConfiguration
LocalConfiguration BE/explicitADmode controls wether
field permissions for backend user groups are additive
(explicitAllow) or subtractive (explicitDeny).

By means of best practice in the IT world, user permissions
should always be configured additive - which is called the
"Principle of Least Privilege" (see https://en.wikipedia.org/wiki/Principle_of_least_privilege).

TYPO3's permission system already uses PoLP in all cases
of backend groups except for the permission system of
"Allowed Content Types"

This is not only important from a security point of view
in terms of permissions (for instance when an extension adds new
CTypes), it's also possible to easily create a hard to
understand mess in explicitDeny mode with nested groups.

The default / fallback in DefaultConfiguration however
has always been explicitDeny since "ever". New instances
always got explicitAllow due to it being set in
FactoryConfiguration - since TYPO3 v6.2.

We can expect the vast majority of instances to run
with explicitAllow already, either since the instance
is younger than ~7 years, or because an admin already
did set the default to explicitAllow for any instance
that works with a non-trivial backend user group setup.

This change turns the default in DefaultConfiguration
around towards explicitAllow. A simple silent upgrade
within install tool sets the value to explicitDeny
in LocalConfiguration for instances that still use this,
so this change is fully transparent and b/w compatible.

Resolves: #94721
Releases: master
Change-Id: Ifd219edce21a1c044541e6c8129134543e8d95ef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70233
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>