[TASK] Deprecate GeneralUtility::minifyJavaScript()
The method "GeneralUtility::minifyJavaScript()"
was introduced back in 2007 when TYPO3 used
"jsmin". Then, the hook was introduced, and
"jsmin" was removed from TYPO3 Core again.

TYPO3 Core also added - in parallel - the concept
of the PageRenderer, which also contains a hook for
extensions (such as EXT:min), or then calls the
ResourceCompressor, which takes care of compressing
and concatenating resources.

The hook is now moved from GeneralUtility and executed
within the ResourceCompressor, so extensions
can still use the hook to compress inline JavaScript.

However, the method is completely mis-placed nowadays
and is moved to the according place in ResourceCompressor.

Resolves: #94791
Releases: master
Change-Id: Ia7b7534ea2fbd44258a191b8a4b16144adaa0c7d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70236
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>